refactor(openapi): extract dynamic-ref helpers into DynamicRefUtils
Move the recursive-type and generic-template dynamic-ref helpers (the
recursive-anchor registry, the generic template-var stack, the $defs
reflection workaround, anchor-name sanitization, and the binding/template
emission logic) out of SchemaDefinitionUtils into a new DynamicRefUtils.

SchemaDefinitionUtils is already very large; this keeps it focused on schema
parsing and groups the opt-in dynamic-refs feature in one place. No behavior
change. populateSchemaProperties widens from private to package-private so
DynamicRefUtils can drive the generic template build.

Addresses review feedback on #2751.
